+++ react/utils/constants.ts
@@ -1,0 +1,1 @@
+export const GIFTS_ADDITIONALINFO_KEY = 'gifts'